+++ src/SourceCode/AlgorithmsFindDeterminand/AlgorithmGauss.java
@@ -1,7 +1,7 @@
 public class AlgorithmGauss implements AlgorithmFindDeterminand {
 
     @Override
-    public int findDeterminand(Matrix matrix) {
+    public double findDeterminand(Matrix matrix) {
         return 0;
     }
 }
